Moved hotkey handling to utilhotkey.h/cpp
