+++ Client/qtTeamTalk/mainwindow.h
@@ -35,6 +35,7 @@
 #include "common.h"
 #include "utilsound.h"
 #include "utilui.h"
+#include "utilhotkey.h"
 
 #if defined(Q_OS_WIN32)
 #define USE_POLL 0
